refactor: put pages into correct folder
