+++ src/components/Router/index.js
@@ -2,10 +2,10 @@
 import MainPage from 'components/pages/MainPage/MainPage';
 import MainPageGame from 'components/pages/MainPage/MainPageGame';
 import Router from 'components/Router/Router';
-import StatisticsPage from 'components/statisticPage/StatisticsPage';
+import StatisticsPage from 'components/pages/statisticPage/StatisticsPage';
 import promoPage from 'components/promo-page/promo-page';
-import authorizationPage from 'components/authorization-page/autorizationPage';
-import teamPage from 'components/team-page/team-page';
+import authorizationPage from 'components/pages/authorization-page/autorizationPage';
+import teamPage from 'components/pages/team-page/team-page';
 import GamesPage from 'components/mini-games-page/MiniGamesPage';
 import AppSprint from 'components/sprint-game/app';
 import PhraseWizard from 'components/phrase-wizard-page/PhraseWizardGame';
