+++ src/components/pages/authorization-page/toast.html
