+++ src/components/pages/statisticPage/GameChart.js
