+++ src/stylesheets/statisticPage/layouts/_gameStatistic.scss
@@ -54,7 +54,7 @@
         li {
             font-size: 1 rem;
             align-self: center;
-            margin: 5px;
+            padding: 5px;
             font-weight: normal;
             text-align: center;
         }
